feat(odbbackup): add tool for backup and restore
